Refactor to get rid of json import and file names
